Extract orbit-local-storage package from orbit-common.
