--- build-support/globalize-orbit-common-local-storage.js
@@ -1,1 +1,0 @@
-window.OC.LocalStorageSource = requireModule("orbit-common/local-storage-source")["default"];
