+++ build-support/globalize-orbit-local-storage.js
@@ -1,0 +1,1 @@
+window.OrbitLocalStorage = requireModule("orbit-local-storage");
