+++ src/orbit-local-storage.js
@@ -1,0 +1,3 @@
+import LocalStorageSource from 'orbit-local-storage/local-storage-source';
+
+export { LocalStorageSource };
