+++ src/orbit-local-storage/lib/queries.js
